+++ wisely/pubspec.yaml
@@ -1,7 +1,7 @@
 name: wisely
 description: A Smart Journal.
 publish_to: 'none'
-version: 0.1.73+88
+version: 0.1.74+89
 
 environment:
   sdk: ">=2.14.0 <3.0.0"
